# Prevent infinite looping in GetString on certain malformed strings